+++ modules/template/helpers/PagePermissionHelper.php
@@ -1,12 +1,12 @@
 <?php
 
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
+/**
+ * @link https://www.humhub.org/
+ * @copyright Copyright (c) HumHub GmbH & Co. KG
+ * @license https://www.humhub.com/licences
  */
 
-namespace humhub\modules\custom_pages\modules\template\models;
+namespace humhub\modules\custom_pages\modules\template\helpers;
 
 use humhub\modules\admin\permissions\ManageModules;
 use humhub\modules\content\helpers\ContentContainerHelper;
@@ -14,12 +14,9 @@ use humhub\modules\custom_pages\permissions\ManagePages;
 use humhub\modules\space\models\Space;
 use Yii;
 
-/**
- * This is the model class for table "custom_pages_template".
- */
-class PagePermission
+class PagePermissionHelper
 {
-    public static function canEdit()
+    public static function canEdit(): bool
     {
         if (Yii::$app->user->isGuest) {
             return false;
@@ -33,9 +30,8 @@ class PagePermission
         return Yii::$app->user->isAdmin() || Yii::$app->user->can([ManageModules::class, ManagePages::class]);
     }
 
-    public static function canTemplate()
+    public static function canTemplate(): bool
     {
         return !Yii::$app->user->isGuest && Yii::$app->user->isAdmin();
     }
-
 }
